requirements.txt
py-cord
yt-dlp>=2025.06.30
python-dotenv
PyNaCl
